requirements.txt
pyenchant=3.2.2
streamlit==1.7.0
streamlit-authenticator==0.1.1
